[#52] 구조체 추가
